Move merged resolvers to separate file
